requirements.txt
torch
accelerate==0.15.0
aiofiles==22.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
anyio==3.6.2
async-timeout==4.0.2
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cos-python-sdk-v5==1.9.22
crcmod==1.7
cycler==0.11.0
diffusers==0.11.1
entrypoints==0.4
et-xmlfile==1.1.0
fastapi==0.89.1
ffmpy==0.3.0
filelock==3.9.0
Flask==2.2.2
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gradio==3.16.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==5.2.0
importlib-resources==5.10.2
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==1.0.3
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.6.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.1
openai==0.25.0
openpyxl==3.0.10
orjson==3.8.5
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
psutil==5.9.4
pycryptodome==3.16.0
pydantic==1.10.4
pydub==0.25.1
pyparsing==3.0.9
pyrsistent==0.19.3
python-dateutil==2.8.2
python-multipart==0.0.5
pytz==2022.7
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tokenizers==0.13.2
toolz==0.12.0
tqdm==4.64.1
transformers==4.25.1
types-pytz==2022.7.0.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.13
uvicorn==0.20.0
websockets==10.4
Werkzeug==2.2.2
xmltodict==0.13.0
yarl==1.8.2
zipp==3.11.0
pymongo
waitress
